Moved split_maybe_args tests into separate tests.rs to satisfy tidy check
